+++ internal/sessiontest/revocation_test.go
@@ -30,6 +30,7 @@ import (
 var (
 	revocationDbType, revocationDbStr = "postgres", "host=127.0.0.1 port=5432 user=testuser dbname=test password='testpassword' sslmode=disable"
 	//revocationDbType, revocationDbStr = "mysql", "testuser:testpassword@tcp(127.0.0.1)/test"
+	//revocationDbType, revocationDbStr = "mssql", "sqlserver://testuser:test-Password@127.0.0.1:1433?database=test"
 
 	revocationPkCounter uint = 2
 )